pyproject.toml
[project]
name = "cov-vina"
version = "0.1.0"
description = "Reference-guided ligand pose prediction with MCS alignment, Vina scoring, and torsion optimization"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "rdkit>=2025.9.5",
    "scipy>=1.17.1",
    "torch>=2.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
